// front/src/features/education/types/IEducation.ts
export interface IEducation {
  finalEdu?: string;
  name?: string;
  state?: string;
  enterDate?: string;
  graduateDate?: string;
  passDate?: string;
  region?: string;
  isQualificationExam?: boolean;
  major?: string;
  universityType?: string;
  universityTime?: string;
  extraMajor?: string;
  extraMajorType?: string;
  grade?: string;
  standardGrade?: string;
  isTransfer?: boolean;
}